pyproject.toml
[tool.poetry]
name = "benzaiten"
version = "0.1.0"
description = ""
authors = ["konumaru <konumaru1022@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.16"
torch = "^1.13.1"
joblib = "^1.2.0"
midi2audio = "^0.1.1"
hydra-core = "^1.3.1"
numpy = "^1.24.0"
scipy = "^1.9.3"
matplotlib = "^3.6.2"
mido = "^1.2.10"
music21 = "^8.1.0"
rich = "^12.6.0"
jupyterlab = "^3.5.2"
pytorch-lightning = "^1.8.6"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"
autoflake = "^2.0.0"
pytest = "^7.2.0"
autopep8 = "^2.0.1"
mypy = "^0.991"
pyproject-flake8 = "^6.0.0.post1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = false
warn_redundant_casts = true

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true

[tool.flake8]
max-line-length = 79
extend-ignore = "E203,"

[tool.pytest.ini_options]
testpaths = ["tests",]
